refactor method from first commit
