+++ app/Rabousy/controllers/Databases.php
@@ -1,18 +1,21 @@
 <?php
-namespace app\controllers;
+
+namespace Rabousy\Controllers;
+
+use Core\Controller;
 
 /**
- * Description of Databases
- *
- * @author christophe
+ * Class Databases
+ * @package Rabousy\Controllers
+ * @author crazafimahatratra
  */
-class Databases extends \core\Controller {
+class Databases extends Controller {
     /**
      *
      * @var \core\Db
      */
     var $db;
-    
+
     public function __construct() {
         parent::__construct();
         $server = $this->session_get("server");
@@ -20,17 +23,17 @@ class Databases extends \core\Controller {
         $password = $this->session_get("password");
         $this->db = new \core\Db($server, $username, $password);
     }
-    
+
     public function get_dbs() {
         $dbs = $this->db->get_databases();
         $this->json_response(array('databases' => $dbs));
     }
-    
+
     public function get_tables($db_name) {
         $tables = $this->db->get_tables($db_name);
         $this->json_response(array('tables' => $tables));
     }
-    
+
     public function get_columns($db_name, $table) {
         $columns = $this->db->get_columns($db_name, $table);
         $res = array();
