+++ composer.json
@@ -18,7 +18,8 @@
     },
     "autoload": {
         "psr-4": {
-            "Rabousy\\": "app/Rabousy"
+            "Rabousy\\": "app/Rabousy",
+            "Core\\": "src/Core"
         }
     }
 }
